+++ Demo/Demo/DemoSlideshow.swift
@@ -1,0 +1,25 @@
+//
+//  DemoSlideshow.swift
+//  Demo
+//
+//  Created by Daniel Saidi on 2024-05-28.
+//  Copyright © 2024 Daniel Saidi. All rights reserved.
+//
+
+import OnboardingKit
+import SwiftUI
+
+struct DemoSlideshow: View {
+
+    let onboarding: LocalizedOnboardingFlow
+
+    @Binding var index: Int
+
+    var body: some View {
+        OnboardingS
+    }
+}
+
+#Preview {
+    DemoSlideshow()
+}