+++ Demo/Demo/Onboarding+Demo.swift
@@ -12,35 +12,38 @@ import OnboardingKit
 extension Onboarding {
     
     static var demoOnboardings: [Onboarding] = [
-        demoListConditional(),
-        demoListCorrectBehavior,
-        demoListDelayed,
-        demoListPopover,
-        demoListSheet,
+        demoConditional(),
+        demoCorrectBehavior,
+        demoDelayed,
+        demoPopover,
+        demoSheet,
         demoToolbarPopover
     ]
     
-    static func demoListConditional(
+    static func demoConditional(
         _ condition: @escaping () -> Bool = { true }
     ) -> ConditionalOnboarding {
         .init(id: "demo.list.conditional", condition: condition)
     }
     
-    static let demoListCorrectBehavior = CorrectBehaviorOnboarding(
+    static let demoCorrectBehavior = CorrectBehaviorOnboarding(
         id: "demo.list.correctbehavior",
         requiredIncorrectAttempts: 3
     )
         
-    static let demoListDelayed = DelayedOnboarding(
+    static let demoDelayed = DelayedOnboarding(
         id: "demo.list.delayed",
         requiredPresentationAttempts: 3
     )
     
-    static let demoListPopover = Onboarding(id: "demo.list.popover")
+    static let demoPopover = Onboarding(id: "demo.list.popover")
     
-    static let demoListSheet = Onboarding(id: "demo.list.sheet")
-    
+    static let demoSheet = Onboarding(id: "demo.list.sheet")
+
     static let demoToolbarPopover = Onboarding(id: "demo.toolbar.popover")
-    
-    static let demoLocalizedFlow = LocalizedOnboardingFlow(id: "demo-flow")
+}
+
+extension LocalizedOnboardingFlow {
+
+    static let demo = LocalizedOnboardingFlow(id: "demo-flow")
 }